test: ChessGameTest로 TC 이동
